+++ lc-lib/transports/tcp/transporttcp.go
@@ -85,8 +85,13 @@ func (t *transportTCP) ReloadConfig(netConfig *transports.Config) bool {
 	if netConfig.Timeout != t.netConfig.Timeout {
 		return true
 	}
-
-	if !reflect.DeepEqual(newConfig.certificate.Certificate, t.config.certificate.Certificate) {
+	if newConfig.certificate != nil && t.config.certificate != nil {
+		if !reflect.DeepEqual(newConfig.certificate.Certificate, t.config.certificate.Certificate) {
+			return true
+		}
+	} else if newConfig.certificate != nil {
+		return true
+	} else if t.config.certificate != nil {
 		return true
 	}
 	if len(newConfig.caList) != len(t.config.caList) {